Move Queuenames in with the celery code, revamp config to allow move to celery 4.x
