+++ app/celery/statistics_tasks.py
@@ -10,7 +10,7 @@ from app.dao.statistics_dao import (
 )
 from app.dao.notifications_dao import get_notification_by_id
 from app.models import NOTIFICATION_STATUS_TYPES_COMPLETED
-from app.config import QueueNames
+from app.celery import QueueNames
 
 
 def create_initial_notification_statistic_tasks(notification):
