+++ app/celery/tasks.py
@@ -19,8 +19,8 @@ from app import (
 )
 from app.aws import s3
 from app.celery import provider_tasks
-from app.config import QueueNames
 from app.dao.inbound_sms_dao import dao_get_inbound_sms_by_id
+from app.celery import QueueNames
 from app.dao.jobs_dao import (
     dao_update_job,
     dao_get_job_by_id,
