+++ app/config.py
@@ -1,9 +1,5 @@
-from datetime import timedelta
 import os
 import json
-
-from celery.schedules import crontab
-from kombu import Exchange, Queue
 
 from app.models import (
     EMAIL_TYPE, SMS_TYPE, LETTER_TYPE,
@@ -16,36 +12,6 @@ if os.environ.get('VCAP_SERVICES'):
     from app.cloudfoundry_config import extract_cloudfoundry_config
 
     extract_cloudfoundry_config()
-
-
-
-class QueueNames(object):
-    PERIODIC = 'periodic-tasks'
-    PRIORITY = 'priority-tasks'
-    DATABASE = 'database-tasks'
-    SEND = 'send-tasks'
-    RESEARCH_MODE = 'research-mode-tasks'
-    STATISTICS = 'statistics-tasks'
-    JOBS = 'job-tasks'
-    RETRY = 'retry-tasks'
-    NOTIFY = 'notify-internal-tasks'
-    PROCESS_FTP = 'process-ftp-tasks'
-
-    @staticmethod
-    def all_queues():
-        return [
-            QueueNames.PRIORITY,
-            QueueNames.PERIODIC,
-            QueueNames.DATABASE,
-            QueueNames.SEND,
-            QueueNames.RESEARCH_MODE,
-            QueueNames.STATISTICS,
-            QueueNames.JOBS,
-            QueueNames.RETRY,
-            QueueNames.NOTIFY,
-            QueueNames.PROCESS_FTP
-        ]
-
 
 
 class Config(object):
